[REF] viewModel 내부 uiState Data class 별도 파일 분리
