+++ app/src/main/java/com/example/letssopt/ui/login/LoginUiState.kt
@@ -1,0 +1,6 @@
+package com.example.letssopt.ui.login
+
+data class LoginUiState(
+        val textId : String = "",
+        val textPw : String = ""
+    )
