+++ app/src/main/java/com/example/letssopt/ui/signup/SignUpUiState.kt
@@ -1,0 +1,7 @@
+package com.example.letssopt.ui.signup
+
+data class SignUpUiState(
+        val textId : String = "",
+        val textPw : String = "",
+        val textCkPw : String = ""
+    )
