+++ app/src/main/java/com/example/letssopt/ui/signup/SignUpViewModel.kt
@@ -10,11 +10,6 @@ import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.flow.update
 
 class SignUpViewModel(application: Application) : AndroidViewModel(application) {
-    data class SignUpUiState(
-        val textId : String = "",
-        val textPw : String = "",
-        val textCkPw : String = ""
-    )
     // ViewModel 내부에서만 값을 변화시키도록 하는 독립적인(실질적인) 변수 DataClass = private
     private val _uiState = MutableStateFlow(SignUpUiState())
     // 외부에서 접근할 수 있는 변수, 단 실질적으로 변화시킬 순 없기에 _uiState로 참조해주면서 동시에 asStateFlow() 함수로 잠굼
